structured the backend
