+++ backend/utils/security.py
@@ -1,0 +1,7 @@
+from werkzeug.security import generate_password_hash, check_password_hash
+
+def hash_password(password):
+    return generate_password_hash(password)
+
+def verify_password(stored_password, provided_password):
+    return check_password_hash(stored_password, provided_password)
